Extracts functionality to utility and geometry files
